Ensure `osu-queue-score-statistics` caches are purged after difficulty calculation completes
Likely fixes
https://github.com/ppy/osu-queue-score-statistics/issues/342 and
https://github.com/ppy/osu/issues/35162.

The polling mechanism used in `osu-queue-score-statistics` to invalidate
caches of difficulty and scoring attributes relies on scanning for new
rows in `bss_process_queue`. The theory behind this change is that this
can lead to stale data remaining in `osu-queue-score-statistics` caches,
because the insertion of a new row with status zero is *much* too early
- difficulty and scoring attributes will not have been *queued for
computation* at that stage, much less actually computed.

Therefore this change forces an invalidation of the cache by forcibly
inserting an already-processed `bss_process_queue` row on completion of
difficulty / score attribute processing, which will hopefully stop the
caches from holding stale data forever.

This can be probably made smarter if need be by reusing
`bss_process_queue` rows that already likely exist, and using a new
`status` value of 3 to indicate full processing completion, but that
would also require changes to the polling mechanism in
`osu-queue-processor`, as the only thing it currently does is check for
an existence of a new row, without inspecting its status. Adjusting it
to check status too is doable, but more complicated to do correctly
without dropping some rows on the way.

The general mechanism of this failure is unconfirmed, but I have tested
this diff to force cache purges on my local environment.